Refactor common validations into separate class
